Refactor hooking logic to use heap-based stuff
This commit refactors the hooking logic to use heap-based variables for everything, since it is required in order to use detours properly. Additionally, we got rid of the static vs class-based `Hooks` functions. We do this to simplify creating hooks since it is a lot of boilerplate that gets reduced.