Add player state getters and setters
This commit adds getters and setters for the player state that we need to track. There are many of these, and they drive the functionality for the rest of the plugin. Some of these include things like the player's preferences in the menu, and their last thrown projectile states, etc.