Implement methods to get player origin and eye angles
This commit adds methods to get a player's origin and eye angles. However, something notable about this commit is that we are now able to take advantage of calling virtual functions instead of hooking via a detour first. This gives us a lot of flexibility to use offsets packaged in the SDKTools extension. In the future, it would be helpful to also support hooking virtual functions.