Add command to respawn the player
This commit adds a command for testing called `sm_jez_respawn` which will respawn the current player. We will need to be able to do this to support grenade playback.

In order to support calling this method, some refactoring has been made to the callables system to support methods that are scanned by signature instead of offset.