Windows support for hooking and calling projectile create
In order to hook and call the projectile create methods on Windows, we need to accomodate the custom calling conventions that are generated by Link Time Code Generation with MSVC. This commit adds support for this, by calling the projectile create methods manually using inline assembly. It also uses some inline assembly at the top of the callback of our hooks to pull out the ECX and EDX registers, which contain the values for `origin` and `angle` respectively.