Add a hook for `CMolotovProjectile::Detonate`
The only grenade type that does not notify us via the `IGameEventsManager2` interface is the molotov/incendiary grenade. Although they technically emit the `molotov_detonate` event, they do not actually inform the server with the entity index, so it is impossible to tell if it was a molotov or an incendiary. This hook that we have added in this commit will happen as the grenade is detonating, allowing us to determine if it is a molotov or incendiary without relying on the event.