Add hook for `OnRunCmd` and `CCSPlayerCanAcquire`
When a player attempts to change weapons, they issue a `CUserCmd` which we are hooking to allow us to fix an issue where players can not switch to a molotov if they have both a molotov and incendiary in their inventory. Currently, we do not apply any resolution to this problem, but we do have the hook in place.

The `CCSPlayerCanAcquire` hook is used to allow players to hold both a molotov and an incendiary at the same time. This allowed us to remove all the options around switching grenade types from the menu.