Use the gamedata offset for `CCSWeaponDataClassName`
This commit changes the hardcoded offset to use the offset that is specified in the game data for obtaining the class name of a `CCSWeaponData` class.

Also, some light refactoring has occurred for the hook of `CCSPlayerCanAcquire` to ensure that we still call the original function to prevent any interested bugs that may occur otherwise. It also ensures that we only attempt to override the server's logic in the case that the original result is 2, which appears to be the case that controls not allowing a pickup of a grenade.