Add feature to switch grenade throw tickrates
This commit allows you to temporarily modify your grenade throw tickrate using the menu. Technically speaking, this feature only works on 128-tick server, since it relies on a trick that was found by @rtldg to modify the start time of a grenade. In a 64-tick server, the start time will be about 7 ticks after the grenade is released. On a 128-tick server, the grenade will be started about 13 ticks after it is released. The result of this precision difference is the reason that jump-throws are different between the two tickrates, since the player is technically in a different position when the grenade is released from their player's hand.