Add the ability to create the short smokegrenade particle
This commit allows us to dispatch a custom particle, which is a short smoke grenade to enable us to have shorter smoke grenade effects. The smoke effect is about 22% as long as the original, which means it passes in just a few seconds rather than 20+ seconds.

We still have to hook up the effect that causes the player's screen to be difficult to see, then we need to replace the standard smoke particle with our custom one.