Add drop-in replacement for the `variant_t` class
This commit adds a drop-in replacement for the `variant_t` class in the SDK, but it doesn't depend on including `cbase.h` which can be a total nightmare.

We will use this in the future to utilize `AcceptInput` on entities (which provides a neat and safe way to delete them, for example).