Remove the smoke screen effect to line it up with the short smoke length
Since we are going to be overwriting normal smokes with the short smoke particle, we need to remove the projectile faster than the normal case. This projectile itself is what causes clients to render the screen effect when you get close, so if you don't remove it the screen effect will last the full length of the normal smoke.