Allow players to switch between their weapons if they have multiple in a slot
This commit allows players to switch between their primary and secondary weapons, even if they have more than one in a given slot.

It isn't perfect, but it should work if the player is using their number keys. It probably won't work as expected if a player is using their scroll wheel.